Made changes to allow file to pass tests
I would like to note that I mistaken a change that was need earlier with the j++ it was in the right spot. In the code I made afew change mostly moving to the use of
functions when ever possible also a few places to clean up code. The most notable change was I got rid of the giant if statement in favor of a function.
This commit also saw the creation of a counter to get the size of the new String... I miss Java `String hello ="bye"; System.out.println(hello);